Working on MAC issues

+++ src/libraries/Calorimeter/CalorimeterDigiHit_factory.cc
@@ -32,7 +32,7 @@ jerror_t CalorimeterDigiHit_factory::brun(jana::JEventLoop *eventLoop, int runnu
 //------------------
 // evnt
 //------------------
-jerror_t CalorimeterDigiHit_factory::evnt(JEventLoop *loop, int eventnumber)
+jerror_t CalorimeterDigiHit_factory::evnt(JEventLoop *loop, uint64_t eventnumber)
 {
 
 	// Code to generate factory data goes here. Add it like:
