+++ src/libraries/ExtVeto/ExtVetoPMTHit_factory.cc
@@ -55,7 +55,7 @@ jerror_t ExtVetoPMTHit_factory::brun(jana::JEventLoop *eventLoop, int runnumber)
 //------------------
 // evnt
 //------------------
-jerror_t ExtVetoPMTHit_factory::evnt(JEventLoop *loop, int eventnumber)
+jerror_t ExtVetoPMTHit_factory::evnt(JEventLoop *loop, uint64_t eventnumber)
 {
 	TranslationTable::ChannelInfo m_channel;
 	TranslationTable::csc_t		  m_csc;
