+++ src/libraries/ctof/marcoCluster_factory.cc
@@ -38,7 +38,7 @@ jerror_t marcoCluster_factory::brun(jana::JEventLoop *eventLoop, int runnumber)
 //------------------
 // evnt
 //------------------
-jerror_t marcoCluster_factory::evnt(JEventLoop *loop, int eventnumber)
+jerror_t marcoCluster_factory::evnt(JEventLoop *loop, uint64_t eventnumber)
 {
