+++ src/libraries/system/BDXEventProcessor.cc
@@ -84,7 +84,7 @@ jerror_t BDXEventProcessor::brun(JEventLoop *eventLoop, int runnumber)
 }
 
 // evnt
-jerror_t BDXEventProcessor::evnt(JEventLoop *loop, int eventnumber)
+jerror_t BDXEventProcessor::evnt(JEventLoop *loop, uint64_t eventnumber)
 {
 	//
 	//	vector<const marcoCluster*> marcoC;
