+++ src/users/andrea/test/JEventProcessor_test.cc
@@ -145,7 +145,7 @@ jerror_t JEventProcessor_test::brun(JEventLoop *loop, int runnumber)
 //------------------
 // evnt
 //------------------
-jerror_t JEventProcessor_test::evnt(JEventLoop *loop, int eventnumber)
+jerror_t JEventProcessor_test::evnt(JEventLoop *loop, uint64_t eventnumber)
 {
 	// This is called for every event. Use of common resources like writing
 	// to a file or filling a histogram should be mutex protected. Using
